Minimal component, problems with "main" method.

+++ src/components/implementation/initf/simple/initf.c
@@ -6,7 +6,7 @@ __attribute__((weak)) int _binary_init_tar_start = 0;
 __attribute__((weak)) int _binary_init_tar_size  = 0;
 
 struct init_fs_info {
-	int size;
+	long size;
 	char *start;
 } info;
 
@@ -30,5 +30,5 @@ int initf_size(void)
 void cos_init(void)
 {
 	info.start = (char*)&_binary_init_tar_start;
-	info.size  = (int)  &_binary_init_tar_size;
+	info.size  = (unsigned long)  &_binary_init_tar_size;
 }
